requirements.txt
streamlit
apscheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client<4.0.0
xlsxwriter
Pillow
pymongo
git+https://github.com/rthalley/dnspython
fpdf
pypdf2
img2pdf
pdfkit
pandas
openpyxl
reportlab
PDFNetPython3
streamlit>1.11.1
pytest>=7.1.2
folium>=0.13
pytest-playwright
pytest-rerunfailures